prinft.c norme et space

+++ ft_all_prints_dux.c
@@ -6,7 +6,7 @@
 /*   By: nsahloum <nsahloum@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/04 14:38:59 by nsahloum          #+#    #+#             */
-/*   Updated: 2020/07/25 02:47:47 by nsahloum         ###   ########.fr       */
+/*   Updated: 2020/07/25 15:12:01 by nsahloum         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_print_num(va_list argp)
 {
-	int nbr;
+	long int nbr;
 	int zero;
 	int neg;
